refactor(utils): splitting out utilities
