+++ source/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './logging';
+export * from './parseOptions';
 export * from './strings';
 export * from './utils';
